Actors lis page added and extracted MoviesPage from App.js
