+++ ui/package.json
@@ -9,6 +9,7 @@
     "milligram": "^1.4.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^7.13.0",
     "react-scripts": "5.0.1",
     "react-select": "^5.10.2",
     "react-toastify": "^11.0.5",
