+++ ui/src/ActorListItem.js
@@ -1,0 +1,7 @@
+export default function ActorListItem(props) {
+    return (
+        <div>
+            <strong>{props.actor.name} {props.actor.surname}</strong>
+        </div>
+    );
+}
